fixup! refactor: Move merge function to common module
