+++ common/pkg/capi/clustertopology/handlers/mutation/meta.go
@@ -5,12 +5,9 @@ package mutation
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"maps"
 	"sync"
 
-	"dario.cat/mergo"
 	"github.com/samber/lo"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -22,6 +19,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/common/pkg/capi/clustertopology/handlers"
+	"github.com/nutanix-cloud-native/cluster-api-runtime-extensions-nutanix/common/pkg/capi/clustertopology/variables"
 )
 
 type MutateFunc func(
@@ -130,7 +128,7 @@ func (mgp metaGeneratePatches) GeneratePatches(
 			// Merge the global variables to the current resource vars. This allows the handlers to access
 			// the variables defined in the cluster class or cluster configuration and use these correctly when
 			// overrides are specified on machine deployment or control plane configuration.
-			mergedVars, err := mergeVariableOverridesWithGlobal(vars, globalVars)
+			mergedVars, err := variables.MergeVariableOverridesWithGlobal(vars, globalVars)
 			if err != nil {
 				log.Error(err, "Failed to merge global variables")
 				return err
@@ -160,57 +158,4 @@ func (mgp metaGeneratePatches) GeneratePatches(
 			return nil
 		},
 	)
-}
-
-// mergeVariableOverridesWithGlobal merges the provided variable overrides with the global variables.
-// It performs a deep merge, ensuring that if a variable exists in both maps, the value from the overrides is used.
-func mergeVariableOverridesWithGlobal(
-	overrideVars, globalVars map[string]apiextensionsv1.JSON,
-) (map[string]apiextensionsv1.JSON, error) {
-	mergedVars := maps.Clone(overrideVars)
-
-	for k, v := range globalVars {
-		// If the value of v is nil, skip it.
-		if v.Raw == nil {
-			continue
-		}
-
-		existingValue, exists := mergedVars[k]
-
-		// If the variable does not exist in the mergedVars or the value is nil, add it and continue.
-		if !exists || existingValue.Raw == nil {
-			mergedVars[k] = v
-			continue
-		}
-
-		// Wrap the value in a temporary key to ensure we can unmarshal to a map.
-		// This is necessary because the values could be scalars.
-		tempValJSON := fmt.Sprintf(`{"value": %s}`, string(existingValue.Raw))
-		tempGlobalValJSON := fmt.Sprintf(`{"value": %s}`, string(v.Raw))
-
-		// Unmarshal the existing value and the global value into maps.
-		var val, globalVal map[string]interface{}
-		if err := json.Unmarshal([]byte(tempValJSON), &val); err != nil {
-			return nil, fmt.Errorf("failed to unmarshal existing value for key %q: %w", k, err)
-		}
-
-		if err := json.Unmarshal([]byte(tempGlobalValJSON), &globalVal); err != nil {
-			return nil, fmt.Errorf("failed to unmarshal global value for key %q: %w", k, err)
-		}
-
-		// Now use mergo to perform a deep merge of the values, retaining the values in `val` if present.
-		if err := mergo.Merge(&val, globalVal); err != nil {
-			return nil, fmt.Errorf("failed to merge values for key %q: %w", k, err)
-		}
-
-		// Marshal the merged value back to JSON.
-		mergedVal, err := json.Marshal(val["value"])
-		if err != nil {
-			return nil, fmt.Errorf("failed to marshal merged value for key %q: %w", k, err)
-		}
-
-		mergedVars[k] = apiextensionsv1.JSON{Raw: mergedVal}
-	}
-
-	return mergedVars, nil
 }
